TPASPKT-463 Remove deprecated tests/data/dpres_mimetypes.json
The mimetypes list is not needed in the tests anymore, since now the
dpres_siptools_research library itself contains the list.

dpres_siptools_research used to have the path to the file configured in
its configuration file and not maintain the list itself. The path was
removed from the test_config fixture in commit
88e0dbefdae0ebd58ca727b350b17d25027f4b96 but apparently the mimetype
list was left behind.

Remove the list since no tests seem to use it.