TPASPKT-1406 Re-enable upload-rest-api E2E tests
upload-rest-api is now available for Metax V3, so add the tests back
with the necessary changes. Most importantly we don't need to deal with
directory identifiers which makes the test setup a bit shorter.

A new dataset and directory for test files is created on Metax and
upload-rest-api respectively for each test run. This also makes
investigating failed test runs easier.